+++ __tests__/shared/components/ScalableRect.jsx
@@ -3,7 +3,11 @@ import React from 'react';
 import { snapshot } from 'utils/jest';
 
 test('Snapshots match', () => {
-  snapshot(<ScalableRect>CONTENT</ScalableRect>);
+  snapshot((
+    <ScalableRect>
+      CONTENT
+    </ScalableRect>
+  ));
   snapshot((
     <ScalableRect
       className="CLASS_NAME"