+++ src/shared/components/ScalableRect/index.jsx
@@ -30,11 +30,15 @@ export default function ScalableRect({ children, className, ratio }) {
       style={{ paddingBottom }}
       styleName="container"
     >
-      <div styleName="wrapper">{children}</div>
+      <div styleName="wrapper">
+        {children}
+      </div>
     </div>
   );
   return className ? (
-    <div className={className}>{rect}</div>
+    <div className={className}>
+      {rect}
+    </div>
   ) : rect;
 }