+++ src/shared/utils/webpack/__mocks__/require.js
@@ -4,5 +4,6 @@ import path from 'path';
 const ROOT = path.resolve(__dirname, '../../../..');
 
 /* eslint-disable global-require, import/no-dynamic-require */
-module.exports = modulePath =>
-  require(modulePath.replace(/^topcoder-react-utils\/dist\/dev/, ROOT));
+module.exports = modulePath => require(
+  modulePath.replace(/^topcoder-react-utils\/dist\/dev/, ROOT),
+);